+++ mybatis-flex-codegen/src/main/java/com/mybatisflex/codegen/config/ColumnConfig.java
@@ -216,6 +216,9 @@ public class ColumnConfig implements Serializable {
     public void setTenantId(Boolean tenantId) {
         this.tenantId = tenantId;
     }
+    public Class<?> getPropertyType() {
+        return propertyType;
+    }
 
     public String getPropertyType() {
         return propertyType;