+++ mybatis-flex-codegen/src/main/java/com/mybatisflex/codegen/entity/Table.java
@@ -131,6 +131,12 @@ public class Table {
 
     public void addColumn(Column column) {
 
+        //排除忽略列
+        if (globalConfig.getStrategyConfig().getIgnoreColumns() != null &&
+            globalConfig.getStrategyConfig().getIgnoreColumns().contains(column.getName().toLowerCase())) {
+            return;
+        }
+
         //主键
         if (primaryKeys != null && primaryKeys.contains(column.getName())) {
             column.setPrimaryKey(true);
@@ -239,7 +245,7 @@ public class Table {
                 tableAnnotation.append(", schema = \"").append(tableConfig.getSchema()).append("\"");
             }
             if (tableConfig.getCamelToUnderline() != null) {
-                tableAnnotation.append(", camelToUnderline = \"").append(tableConfig.getCamelToUnderline()).append("\"");
+                tableAnnotation.append(", camelToUnderline = ").append(tableConfig.getCamelToUnderline());
             }
             if (tableConfig.getInsertListenerClass() != null) {
                 tableAnnotation.append(", onInsert = ").append(tableConfig.getInsertListenerClass().getSimpleName()).append(".class");