+++ mybatis-flex-core/src/main/java/com/mybatisflex/core/query/QueryChain.java
@@ -29,28 +29,16 @@ import java.util.Optional;
  * @author 王帅
  * @since 2023-07-22
  */
-public class QueryWrapperChain<T> extends QueryWrapperAdapter<QueryWrapperChain<T>> {
+public class QueryChain<T> extends QueryWrapperAdapter<QueryChain<T>> {
 
     private final BaseMapper<T> baseMapper;
 
-    public QueryWrapperChain(BaseMapper<T> baseMapper) {
+    public QueryChain(BaseMapper<T> baseMapper) {
         this.baseMapper = baseMapper;
     }
 
-    public static <E> QueryWrapperChain<E> create(BaseMapper<E> baseMapper) {
-        return new QueryWrapperChain<>(baseMapper);
-    }
-
-    public boolean remove() {
-        return SqlUtil.toBool(baseMapper.deleteByQuery(this));
-    }
-
-    public boolean update(T entity) {
-        return SqlUtil.toBool(baseMapper.updateByQuery(entity, this));
-    }
-
-    public boolean update(T entity, boolean ignoreNulls) {
-        return SqlUtil.toBool(baseMapper.updateByQuery(entity, ignoreNulls, this));
+    public static <E> QueryChain<E> of(BaseMapper<E> baseMapper) {
+        return new QueryChain<>(baseMapper);
     }
 
     public long count() {